Minor fixes in places; DefaultQCH Java test-inputs and test-program are now able to handle arrays. Needs further testing.

+++ Plugins/DefaultQuestionCriteriaHandlers/src/pals/plugins/handlers/defaultqch/java/Utils.java
@@ -63,10 +63,11 @@ public class Utils
     }
     /**
      * @param fullName The full name of the class to parse.
+     * @param cl A class-loader to also be used to find a class; can be null.
      * @return The class of the full-name; supports primitives simply by
      * their name, as well as String by its short name.
      */
-    public static Class parseClass(String fullName)
+    public static Class parseClass(String fullName, ClassLoader cl)
     {
         if(fullName == null)
             return null;
@@ -74,25 +75,54 @@ public class Utils
         {
             case "byte":
                 return byte.class;
+            case "byte[]":
+                return byte[].class;
             case "short":
                 return short.class;
+            case "short[]":
+                return short[].class;
             case "int":
                 return int.class;
+            case "int[]":
+                return int[].class;
             case "long":
                 return long.class;
+            case "long[]":
+                return long[].class;
             case "float":
                 return float.class;
+            case "float[]":
+                return float[].class;
             case "double":
                 return double.class;
+            case "double[]":
+                return double[].class;
             case "boolean":
             case "bool":
                 return boolean.class;
+            case "boolean[]":
+            case "bool[]":
+                return boolean[].class;
             case "char":
                 return char.class;
+            case "char[]":
+                return char[].class;
             case "String":
             case "string":
                 return String.class;
+            case "String[]":
+            case "string[]":
+                return String[].class;
             default:
+                // Try custom class-loader
+                try
+                {
+                    return Class.forName(fullName, true, cl);
+                }
+                catch(ClassNotFoundException ex)
+                {
+                }
+                // Try the runtime class-loader
                 try
                 {
                     return Class.forName(fullName);
@@ -122,14 +152,20 @@ public class Utils
         try
         {
             File f = new File(pathTemp);
-            // Purge .class files
-            File[] t = Files.getAllFiles(pathTemp, false, true, ".class", true);
-            for(File c : t)
-                c.delete();
-            // Remove empty directories
-            t = Files.getDirsEmpty(pathTemp);
-            for(File c : t)
-                c.delete();
+            // Ensure dir exists
+            if(!f.exists())
+                f.mkdir();
+            else
+            {
+                // Purge .class files
+                File[] t = Files.getAllFiles(pathTemp, false, true, ".class", true);
+                for(File c : t)
+                    c.delete();
+                // Remove empty directories
+                t = Files.getDirsEmpty(pathTemp);
+                for(File c : t)
+                    c.delete();
+            }
         }
         catch(IOException ex)
         {
@@ -173,17 +209,13 @@ public class Utils
      * @param method The method to invoke.
      * @param whiteListedClasses An array of white-listed classes.
      * @param outputValue Indicates if to output the value from the method invoked.
-     * @param inputTypes Array of types, corresponding with the indexes of inputs.
-     * @param inputs The input values, to be parsed, for locating and invoking the method.
+     * @param inputsArgs Array of input arguments for the Java sandbox.
      * @return Arguments for launching sandbox program.
      */
-    public static String[] buildJavaSandboxArgs(NodeCore core, String javaSandboxPath, String directory, String className, String method, String[] whiteListedClasses, boolean outputValue, String[] inputTypes, String[] inputs)
+    public static String[] buildJavaSandboxArgs(NodeCore core, String javaSandboxPath, String directory, String className, String method, String[] whiteListedClasses, boolean outputValue, String[] inputsArgs)
     {
-        if(inputTypes.length != inputs.length)
-            throw new IllegalArgumentException();
-        
         final int BASE_ARGS = 8;
-        String[] buffer = new String[BASE_ARGS+inputs.length];
+        String[] buffer = new String[BASE_ARGS+inputsArgs.length];
         // Setup base args
         buffer[0] = "-jar";
         buffer[1] = PalsProcess.formatPath(javaSandboxPath);
@@ -194,16 +226,48 @@ public class Utils
         buffer[6] = outputValue ? "1" : "0";
         buffer[7] = Integer.toString(core.getSettings().getInt("tools/javasandbox/timeout_ms", 10000));
         // Setup input args
-        for(int i = 0; i < inputs.length; i++)
-            buffer[BASE_ARGS+i] = inputTypes[i]+"="+inputs[i];
+        for(int i = 0; i < inputsArgs.length; i++)
+            buffer[BASE_ARGS+i] = inputsArgs[i];
         return buffer;
     }
-    public static void formatInputs(NodeCore core, String[] inputTypes, String[] inputRow)
+    /**
+     * Formats the specified input-arguments, allowing for random numbers etc.
+     * 
+     * @param core Current instance of the core.
+     * @param inputArgs The input arguments, for the Java sandbox, to be
+     * formatted.
+     */
+    public static void formatInputs(NodeCore core, String[] inputArgs)
     {
-        for(int i = 0; i < inputRow.length; i++)
+        int             ind;        // Index of =
+        String          k;          // Key for argument.
+        String[]        v;          // Value(s) for an argument.
+        String          t;          // Specific value from v.
+        StringBuilder   buffer;     // Used to rebuild args.
+        for(int i = 0; i < inputArgs.length; i++)
         {
-            if(inputRow[i].startsWith("rand"))
-                inputRow[i] = formatInputs_inputRandom(core, inputTypes[i], inputRow[i]);
+            // Fetch entire argument
+            k = inputArgs[i];
+            // Fetch values for argument; may be array, so split by ,
+            ind = k.indexOf('=');
+            v = ind == k.length()-1 ? new String[0] : inputArgs[i].substring(ind+1).split(",");
+            // Fetch key
+            k = k.substring(0, ind);
+            // Iterate and format each value; we will also rebuild the arg at the same time
+            buffer = new StringBuilder();
+            buffer.append(k).append("=");
+            for(int j = 0; j < v.length; j++)
+            {
+                t = v[j];
+                // Format current value for random value
+                if(t.startsWith("rand("))
+                    t = formatInputs_inputRandom(core, k, t);
+                buffer.append(t).append(",");
+            }
+            // Rebuild argument
+            if(v.length > 0)
+                buffer.deleteCharAt(buffer.length()-1);
+            inputArgs[i] = buffer.toString();
         }
     }
     private static String formatInputs_inputRandom(NodeCore core, String type, String input)
@@ -227,22 +291,32 @@ public class Utils
             switch(type)
             {
                 case "byte":
+                case "byte:arr":
                     return Byte.toString((byte)rv);
                 case "short":
+                case "short:arr":
                     return Short.toString((short)rv);
                 case "integer":
+                case "integer:arr":
                 case "int":
+                case "int:arr":
                     return Integer.toString(rv);
                 case "long":
+                case "long:arr":
                     return Long.toString((long)rv);
                 case "float":
+                case "float:arr":
                     return Float.toString((float)rv);
                 case "double":
+                case "double:arr":
                     return Double.toString((double)rv);
                 case "bool":
+                case "bool:arr":
                 case "boolean":
+                case "boolean:arr":
                     return Boolean.toString(rv % 2 == 1);
                 case "char":
+                case "char:arr":
                     return Character.toString((char)rv);
             }
         }
